requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.2.0
redis==5.0.1
django-redis==5.4.0
paypalrestsdk==1.13.1
django-filter==23.5
django-storages==1.14.2
django-allauth==0.60.1
argon2-cffi==23.1.0
gunicorn==21.2.0
celery==5.3.6
django-celery-beat==2.5.0
django-celery-results==2.5.1 